Reworked Armor- and WeaponComponent Reworked CombatSystem to factor in weapon/armor/resistances

+++ TheAlchemist/Components/ArmorComponent.cs
@@ -8,7 +8,7 @@ namespace TheAlchemist.Components
 {
     class ArmorComponent : Component<ArmorComponent>
     {
-        public float FlatMitigation { get; set; }
-        public float PercentMitigation { get; set; }
+        public int FlatMitigation { get; set; }
+        public int PercentMitigation { get; set; }
     }
 }
